access and refreh token

+++ src/utils/cloudinary.js
@@ -1,7 +1,6 @@
 import {v2 as cloudinary} from "cloudinary"
 import fs from "fs"
 
-
 cloudinary.config({ 
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
   api_key: process.env.CLOUDINARY_API_KEY, 
@@ -19,9 +18,16 @@ const uploadOnCloudinary = async (localFilePath) => {
         //file uploaded successfully
 
         console.log("file is uploaded on cloudinary", response.url);
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+        console.log(response)
         return response;
     } catch (error) {
-        fs.unlinkSync(localFilePath)
+        console.log("cloudinary cant upload", error)
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
         return null;
     }
